requirements.txt
amqp==5.1.0
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.21.42
botocore==1.24.42
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ddtrace==0.59.1
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
idna==3.3
jmespath==1.0.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0

django-utils2~=2.12.1